requirements.txt
alembic==1.5.4
Authlib==0.15.3
cachelib==0.1.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.5
DateTime==4.3
ecdsa==0.14.1
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jose==1.0.0
linecache2==1.0.0
Mako==1.1.4
MarkupSafe==1.1.1
numpy==1.20.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-jose==3.2.0
pytz==2021.1
requests==2.25.1
rsa==4.7.1
six==1.15.0
SQLAlchemy==1.3.23
sqlparse==0.4.1
termcolor==1.1.0
timezonefinder==5.2.0
traceback2==1.4.0
urllib3==1.26.3
Werkzeug==1.0.1
zope.interface==5.2.0
